app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.5
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
cliapp==1.0.9
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Limiter==3.3.0
Flask-Mail==0.9.1
Flask-WTF==1.1.1
frozenlist==1.3.3
ghp-import==2.1.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
limits==3.3.0
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-material==9.1.8
mkdocs-material-extensions==1.1.1
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
ordered-set==4.1.0
packaging==23.0
pathy==0.10.1
preshed==3.0.8
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
Pygments==2.14.0
pymdown-extensions==9.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
sendgrid==6.10.0
sendgrid-python==0.1.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starkbank-ecdsa==2.2.0
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wasabi==1.1.1
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.1
zipp==3.15.0
